pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "imagen"
version = "0.1.0"
description = "Multi-provider image generation CLI (Gemini Nano Banana, OpenAI gpt-image, local ComfyUI)."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Tomoya Ishii" }]
dependencies = [
    "typer>=0.12",
    "google-genai>=1.0",
    "openai>=1.50",
    "httpx>=0.27",
    "jinja2>=3.1",
    "pydantic>=2.6",
    "tomli-w>=1.0",
    "pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-httpx>=0.30",
    "respx>=0.21",
]

[project.scripts]
imagen = "imagen.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/imagen"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
